Fixed classical test
